frontend/frontend_requirements.txt
streamlit
requests
python-dotenv
google-cloud-logging